Use AnyEvent::Handle for queuing results
When sending results back to the parent process, the scheduler now
queues outbound writes with AnyEvent::Handle.  This lets the scheduler
queue as many results as it has while still being able to handle future
work.

With this change, t/many.t spawning 500 threads does the right thing.
Previously that only worked up to 170 threads, or so.

I still get an intermitent problem (once every 30 runs or so) when
t/many.t hangs.  It seems that the parent process sends a job but the
scheduler never receives it.  The scheduler blocks waiting for a new job
and the parent process blocks waiting for the result.  None of the
workers are busy.